+++ extension/src/Repository/index.test.ts
@@ -3,14 +3,14 @@ import { join, resolve } from 'path'
 import { Config } from '../Config'
 import { SourceControlManagement } from './views/SourceControlManagement'
 import { mocked } from 'ts-jest/utils'
-import { DecorationProvider } from '../DecorationProvider'
+import { DecorationProvider } from './DecorationProvider'
 import { Repository, RepositoryState } from '.'
 import { listDvcOnlyRecursive, status } from '../cli/reader'
 import { getAllUntracked } from '../git'
 
 jest.mock('@hediet/std/disposable')
 jest.mock('./views/SourceControlManagement')
-jest.mock('../DecorationProvider')
+jest.mock('./DecorationProvider')
 jest.mock('../cli/reader')
 jest.mock('../git')
